gs-hack-user/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
boto3 = "^1.18.30"
awscli = "^1.20.30"
pytesseract = "^0.3.8"
Pillow = "^8.3.1"
opencv-python = "^4.5.3"
python-multipart = "^0.0.5"
pandas = "^1.3.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
